Add Lunar Sword and Ash of Dawn items and models
Introduces the Lunar Sword item, its crafting recipe, and associated model and texture files. Adds the Ash of Dawn item and its projectile entity, with supporting procedures and JSON definitions. Also adds new armor renderers for Mythril, Mythril Enchanted, and Runite sets. Updates various block, item, and entity files to support these additions and makes minor adjustments across multiple elements for integration.

+++ src/main/java/net/mcreator/losthorizon/procedures/BloodlustProcedureProcedure.java
@@ -39,13 +39,13 @@ public class BloodlustProcedureProcedure {
 						.getEnchantmentLevel(world.registryAccess().lookupOrThrow(Registries.ENCHANTMENT).getOrThrow(ResourceKey.create(Registries.ENCHANTMENT, ResourceLocation.parse("losthorizon:bloodlust")))) != 0
 				&& (sourceentity instanceof LivingEntity _livEnt ? _livEnt.getHealth() : -1) < 20) {
 			if (sourceentity instanceof LivingEntity _entity)
-				_entity.setHealth((float) ((sourceentity instanceof LivingEntity _livEnt ? _livEnt.getHealth() : -1) + 1 * (sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)
-						.getEnchantmentLevel(world.registryAccess().lookupOrThrow(Registries.ENCHANTMENT).getOrThrow(ResourceKey.create(Registries.ENCHANTMENT, ResourceLocation.parse("losthorizon:bloodlust"))))));
+				_entity.setHealth((sourceentity instanceof LivingEntity _livEnt ? _livEnt.getHealth() : -1) + 1 * (sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)
+						.getEnchantmentLevel(world.registryAccess().lookupOrThrow(Registries.ENCHANTMENT).getOrThrow(ResourceKey.create(Registries.ENCHANTMENT, ResourceLocation.parse("losthorizon:bloodlust")))));
 			if (world instanceof ServerLevel _level) {
 				(sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)
 						.hurtAndBreak(
-								(int) (2 * (sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)
-										.getEnchantmentLevel(world.registryAccess().lookupOrThrow(Registries.ENCHANTMENT).getOrThrow(ResourceKey.create(Registries.ENCHANTMENT, ResourceLocation.parse("losthorizon:bloodlust"))))),
+								2 * (sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY)
+										.getEnchantmentLevel(world.registryAccess().lookupOrThrow(Registries.ENCHANTMENT).getOrThrow(ResourceKey.create(Registries.ENCHANTMENT, ResourceLocation.parse("losthorizon:bloodlust")))),
 								_level, null, _stkprov -> {
 								});
 			}
